Product API Integration: I help integrate and link to the front end. New products can now be added via the API. This is part of the Noura section.
Vue component updates: I also modified the Vue components, including ShoppingCart.vue and Checkout.vue, to make sure the new product functionality works smoothly. To accommodate these changes, we've made special fixes in the ‘Aneeqa’ section.

Database update: I was in charge of the database integration and made sure that everything was in line with the new cart functionality.